Splitting out LocalFile and LocalDirectory spec
